refactor(ui): break deployment.tsx into focused components #266
Split the 1109-line deployment.tsx into a thin shell (217 LOC) that
delegates to five extracted sub-components under deployment/:
- firmware-filters.tsx (66 LOC) — status pills + model dropdown
- firmware-list.tsx (247 LOC) — card grid with SoD actions
- vulnerability-tab.tsx (218 LOC) — severity table + pagination
- reports-tab.tsx (204 LOC) — report generation + export
- audit-log-tab.tsx (355 LOC) — sortable table with filters

No UI changes. All files under 400 lines. Build verified.
