+++ src/app/components/deployment/vulnerability-tab.tsx
@@ -1,0 +1,215 @@
+import { ChevronRight, ChevronLeft, Bug } from "lucide-react";
+import { cn } from "../../../lib/utils";
+import { VULN_PAGE_SIZE, SEVERITY_CONFIG, REMEDIATION_STYLES } from "./deployment-constants";
+import type { VulnerabilityEntry, VulnSeverity, RemediationStatus } from "./deployment-types";
+
+interface VulnerabilityTabProps {
+  vulnerabilities: VulnerabilityEntry[];
+  filteredVulnerabilities: VulnerabilityEntry[];
+  paginatedVulnerabilities: VulnerabilityEntry[];
+  totalVulnPages: number;
+  vulnSeverityFilter: VulnSeverity | "All";
+  setVulnSeverityFilter: (value: VulnSeverity | "All") => void;
+  vulnPage: number;
+  setVulnPage: (value: number | ((prev: number) => number)) => void;
+  canManageVulns: boolean;
+  handleRemediationChange: (id: string, status: RemediationStatus) => void;
+}
+
+export function VulnerabilityTab({
+  vulnerabilities,
+  filteredVulnerabilities,
+  paginatedVulnerabilities,
+  totalVulnPages,
+  vulnSeverityFilter,
+  setVulnSeverityFilter,
+  vulnPage,
+  setVulnPage,
+  canManageVulns,
+  handleRemediationChange,
+}: VulnerabilityTabProps) {
+  return (
+    <div className="space-y-3">
+      {/* Severity filter pills -- Story 11.4 AC3 */}
+      <div className="flex items-center gap-1.5">
+        {(["All", "Critical", "High", "Medium", "Low"] as const).map((s) => {
+          const count =
+            s === "All"
+              ? vulnerabilities.length
+              : vulnerabilities.filter((v) => v.severity === s).length;
+          const pillColor =
+            s === "All"
+              ? ""
+              : s === "Critical"
+                ? "bg-red-50 text-red-700 hover:bg-red-100"
+                : s === "High"
+                  ? "bg-orange-50 text-orange-700 hover:bg-orange-100"
+                  : s === "Medium"
+                    ? "bg-amber-50 text-amber-700 hover:bg-amber-100"
+                    : "bg-emerald-50 text-emerald-700 hover:bg-emerald-100";
+
+          return (
+            <button
+              key={s}
+              onClick={() => {
+                setVulnSeverityFilter(s);
+                setVulnPage(1);
+              }}
+              className={cn(
+                "rounded-full px-2.5 py-1 text-[12px] font-medium transition-colors duration-150",
+                vulnSeverityFilter === s
+                  ? "bg-accent text-white"
+                  : s === "All"
+                    ? "bg-muted text-muted-foreground hover:bg-muted/80"
+                    : pillColor,
+              )}
+            >
+              {s} ({count})
+            </button>
+          );
+        })}
+      </div>
+
+      {/* Vulnerability Table -- Story 11.4 */}
+      <div className="overflow-auto rounded-sm border border-border">
+        <table className="w-full text-sm">
+          <caption className="sr-only">Deployment vulnerability list</caption>
+          <thead>
+            <tr className="border-b-2 border-border bg-muted/50">
+              <th
+                scope="col"
+                className="px-3 py-2.5 text-left text-[13px] font-bold text-muted-foreground uppercase tracking-wider"
+              >
+                CVE ID
+              </th>
+              <th
+                scope="col"
+                className="px-3 py-2.5 text-left text-[13px] font-bold text-muted-foreground uppercase tracking-wider"
+              >
+                Severity
+              </th>
+              <th
+                scope="col"
+                className="px-3 py-2.5 text-left text-[13px] font-bold text-muted-foreground uppercase tracking-wider"
+              >
+                Affected Component
+              </th>
+              <th
+                scope="col"
+                className="px-3 py-2.5 text-left text-[13px] font-bold text-muted-foreground uppercase tracking-wider"
+              >
+                Remediation Status
+              </th>
+              <th
+                scope="col"
+                className="px-3 py-2.5 text-left text-[13px] font-bold text-muted-foreground uppercase tracking-wider"
+              >
+                Firmware Version
+              </th>
+            </tr>
+          </thead>
+          <tbody>
+            {paginatedVulnerabilities.length === 0 ? (
+              <tr>
+                <td colSpan={5} className="px-3 py-12 text-center">
+                  <div className="flex flex-col items-center gap-2">
+                    <Bug className="h-8 w-8 text-muted-foreground/30" />
+                    <p className="text-sm text-muted-foreground">No vulnerabilities found</p>
+                  </div>
+                </td>
+              </tr>
+            ) : (
+              paginatedVulnerabilities.map((vuln) => {
+                const sevCfg = SEVERITY_CONFIG[vuln.severity];
+                return (
+                  <tr
+                    key={vuln.id}
+                    className="border-b border-border last:border-0 hover:bg-muted/30 transition-colors duration-150"
+                  >
+                    <td className="px-3 py-2.5 font-mono text-[13px] font-medium text-foreground">
+                      {vuln.cveId}
+                    </td>
+                    <td className="px-3 py-2.5">
+                      <span
+                        className={cn(
+                          "rounded-full px-2 py-0.5 text-[12px] font-semibold",
+                          sevCfg.bg,
+                          sevCfg.text,
+                        )}
+                      >
+                        {vuln.severity}
+                      </span>
+                    </td>
+                    <td className="px-3 py-2.5 text-muted-foreground">{vuln.affectedComponent}</td>
+                    <td className="px-3 py-2.5">
+                      {/* Story 11.5 -- Inline status dropdown for Admin/Manager, badge for Viewer */}
+                      {canManageVulns ? (
+                        <select
+                          value={vuln.remediationStatus}
+                          onChange={(e) =>
+                            handleRemediationChange(vuln.id, e.target.value as RemediationStatus)
+                          }
+                          className={cn(
+                            "rounded px-2 py-0.5 text-[12px] font-medium border-0 cursor-pointer focus:outline-none focus:ring-1 focus:ring-ring",
+                            REMEDIATION_STYLES[vuln.remediationStatus],
+                          )}
+                        >
+                          <option value="Open">Open</option>
+                          <option value="In Progress">In Progress</option>
+                          <option value="Resolved">Resolved</option>
+                        </select>
+                      ) : (
+                        <span
+                          className={cn(
+                            "rounded px-2 py-0.5 text-[12px] font-medium",
+                            REMEDIATION_STYLES[vuln.remediationStatus],
+                          )}
+                        >
+                          {vuln.remediationStatus}
+                        </span>
+                      )}
+                      {vuln.resolvedDate && (
+                        <span className="ml-1.5 text-[12px] text-muted-foreground">
+                          ({vuln.resolvedDate})
+                        </span>
+                      )}
+                    </td>
+                    <td className="px-3 py-2.5 text-muted-foreground">{vuln.firmwareVersion}</td>
+                  </tr>
+                );
+              })
+            )}
+          </tbody>
+        </table>
+      </div>
+
+      {/* Pagination */}
+      {filteredVulnerabilities.length > VULN_PAGE_SIZE && (
+        <div className="flex items-center justify-between text-sm text-muted-foreground">
+          <span>
+            Showing {Math.min((vulnPage - 1) * VULN_PAGE_SIZE + 1, filteredVulnerabilities.length)}
+            {" - "}
+            {Math.min(vulnPage * VULN_PAGE_SIZE, filteredVulnerabilities.length)} of{" "}
+            {filteredVulnerabilities.length} entries
+          </span>
+          <div className="flex items-center gap-1">
+            <button
+              onClick={() => setVulnPage((p) => Math.max(1, p - 1))}
+              disabled={vulnPage <= 1}
+              className="rounded-sm border border-border p-1 hover:bg-muted disabled:opacity-60 disabled:cursor-not-allowed"
+            >
+              <ChevronLeft className="h-3.5 w-3.5" />
+            </button>
+            <button
+              onClick={() => setVulnPage((p) => Math.min(totalVulnPages, p + 1))}
+              disabled={vulnPage >= totalVulnPages}
+              className="rounded-sm border border-border p-1 hover:bg-muted disabled:opacity-60 disabled:cursor-not-allowed"
+            >
+              <ChevronRight className="h-3.5 w-3.5" />
+            </button>
+          </div>
+        </div>
+      )}
+    </div>
+  );
+}
